Move config to conf
