+++ .gitignore
@@ -4,7 +4,7 @@ __pycache__/
 *$py.class
 .sass-cache/
 .venv/
-config/staticfiles
+staticfiles
 BrowserStackLocal
 node_modules/
