+++ core/tests/test_context_processor.py
@@ -1,10 +1,10 @@
-from config import context_processors
+from core import context_processors
 
 
 def test_feature_flags_installed(settings):
     processors = settings.TEMPLATES[0]['OPTIONS']['context_processors']
 
-    assert 'config.context_processors.feature_flags' in processors
+    assert 'core.context_processors.feature_flags' in processors
 
 
 def test_feature_returns_expected_features(settings):
