+++ docker/cmd-webserver.sh
@@ -1,4 +1,4 @@
 #!/bin/bash -xe
 
 python /usr/src/app/manage.py collectstatic --noinput
-gunicorn config.wsgi --bind 0.0.0.0:$PORT --log-file -
+gunicorn conf.wsgi --bind 0.0.0.0:$PORT --log-file -
