+++ docker/env.test.json
@@ -49,7 +49,9 @@
       "HEADER_FOOTER_URLS_SOO",
       "HEADER_FOOTER_URLS_EVENTS",
       "HEADER_FOOTER_URLS_CONTACT_US",
-      "HEADER_FOOTER_URLS_DIT"
+      "HEADER_FOOTER_URLS_DIT",
+      "FEATURE_MAINTENANCE_MODE_ENABLED",
+      "FEATURE_SEARCH_ENGINE_INDEXING_DISABLED"
     ]
   }
 }
