+++ requirements_test.txt
@@ -13,7 +13,7 @@ contextlib2==0.5.5
 coverage==4.5.1           # via codecov, pytest-cov
 directory-api-external==4.2.1
 directory-client-core==0.2.1
-directory-components==2.6.0
+directory-components==2.7.2
 directory-constants==6.8.0
 directory-healthcheck==0.4.2
 directory-sso-api-client==2.5.1
